pyproject.toml
[project]
name = "picam"
version = "0.1.0"
description = ""
authors = [
    {name = "Will",email = "wlwatkins@pm.me"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless (>=4.11.0.86,<5.0.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "escpos-python (>=1.0.2,<2.0.0)",
    "rpi-gpio (>=0.7.1,<0.8.0)"
]

[tool.poetry.scripts]
picam = "picam.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
